+++ python/instrumentation/openinference-instrumentation-langchain/tests/openinference/instrumentation/langchain/test_instrumentor.py
@@ -1,4 +1,5 @@
 import asyncio
+import json
 import logging
 import random
 from contextlib import suppress
@@ -9,12 +10,12 @@ import numpy as np
 import openai
 import pytest
 from httpx import AsyncByteStream, Response, SyncByteStream
-from langchain.chains import RetrievalQA
+from langchain.chains import LLMChain, RetrievalQA
 from langchain_community.embeddings import FakeEmbeddings
 from langchain_community.retrievers import KNNRetriever
+from langchain_core.prompts import PromptTemplate
 from langchain_openai import ChatOpenAI
 from openinference.instrumentation.langchain import LangChainInstrumentor
-from openinference.instrumentation.openai import OpenAIInstrumentor
 from openinference.semconv.trace import (
     DocumentAttributes,
     EmbeddingAttributes,
@@ -26,7 +27,6 @@ from openinference.semconv.trace import (
 )
 from opentelemetry import trace as trace_api
 from opentelemetry.sdk import trace as trace_sdk
-from opentelemetry.sdk.resources import Resource
 from opentelemetry.sdk.trace.export import SimpleSpanProcessor
 from opentelemetry.sdk.trace.export.in_memory_span_exporter import InMemorySpanExporter
 from opentelemetry.semconv.trace import SpanAttributes as OTELSpanAttributes
@@ -54,6 +54,8 @@ def test_callback_llm(
     completion_usage: Dict[str, Any],
 ) -> None:
     question = randstr()
+    template = "{context}{question}"
+    prompt = PromptTemplate(input_variables=["context", "question"], template=template)
     output_messages: List[Dict[str, Any]] = (
         chat_completion_mock_stream[1] if is_stream else [{"role": randstr(), "content": randstr()}]
     )
@@ -81,7 +83,11 @@ def test_callback_llm(
         texts=documents,
         embeddings=FakeEmbeddings(size=2),
     )
-    rqa = RetrievalQA.from_chain_type(llm=chat_model, retriever=retriever)
+    rqa = RetrievalQA.from_chain_type(
+        llm=chat_model,
+        retriever=retriever,
+        chain_type_kwargs={"prompt": prompt},
+    )
     with suppress(openai.BadRequestError):
         if is_async:
             asyncio.run(rqa.ainvoke({"query": question}))
@@ -149,6 +155,15 @@ def test_callback_llm(
     assert llm_attributes.pop(OPENINFERENCE_SPAN_KIND, None) == CHAIN.value
     assert llm_attributes.pop(INPUT_VALUE, None) is not None
     assert llm_attributes.pop(INPUT_MIME_TYPE, None) == JSON.value
+    assert llm_attributes.pop(LLM_PROMPT_TEMPLATE, None) == template
+    assert isinstance(
+        template_variables_json_string := llm_attributes.pop(LLM_PROMPT_TEMPLATE_VARIABLES, None),
+        str,
+    )
+    assert json.loads(template_variables_json_string) == {
+        "context": "\n\n".join(documents),
+        "question": question,
+    }
     if status_code == 200:
         assert llm_attributes.pop(OUTPUT_VALUE, None) == output_messages[0]["content"]
     elif status_code == 400:
@@ -202,18 +217,39 @@ def test_callback_llm(
         ) == "BadRequestError"
     assert oai_attributes == {}
 
-    # The remaining span is from the openai instrumentor.
-    openai_span = spans_by_name.popitem()[1]
-    assert openai_span.parent is not None
-    if is_async:
-        # FIXME: it's unclear why the context fails to propagate.
-        assert openai_span.parent.span_id == llm_span.context.span_id
-        assert openai_span.context.trace_id == llm_span.context.trace_id
-    else:
-        assert openai_span.parent.span_id == oai_span.context.span_id
-        assert openai_span.context.trace_id == oai_span.context.trace_id
-
     assert spans_by_name == {}
+
+
+def test_chain_metadata(
+    respx_mock: MockRouter,
+    in_memory_span_exporter: InMemorySpanExporter,
+    completion_usage: Dict[str, Any],
+) -> None:
+    url = "https://api.openai.com/v1/chat/completions"
+    respx_kwargs: Dict[str, Any] = {
+        "json": {
+            "choices": [
+                {
+                    "index": 0,
+                    "message": {"role": "assistant", "content": "nock nock"},
+                    "finish_reason": "stop",
+                }
+            ],
+            "model": "gpt-3.5-turbo",
+            "usage": completion_usage,
+        }
+    }
+    respx_mock.post(url).mock(return_value=Response(status_code=200, **respx_kwargs))
+    prompt_template = "Tell me a {adjective} joke"
+    prompt = PromptTemplate(input_variables=["adjective"], template=prompt_template)
+    llm = LLMChain(llm=ChatOpenAI(), prompt=prompt, metadata={"category": "jokes"})
+    llm.predict(adjective="funny")
+    spans = in_memory_span_exporter.get_finished_spans()
+    spans_by_name = {span.name: span for span in spans}
+
+    assert (llm_chain_span := spans_by_name.pop("LLMChain")) is not None
+    assert llm_chain_span.attributes
+    assert llm_chain_span.attributes.get("metadata.category") == "jokes"
 
 
 @pytest.fixture
@@ -242,10 +278,8 @@ def in_memory_span_exporter() -> InMemorySpanExporter:
 
 @pytest.fixture(scope="module")
 def tracer_provider(in_memory_span_exporter: InMemorySpanExporter) -> trace_api.TracerProvider:
-    resource = Resource(attributes={})
-    tracer_provider = trace_sdk.TracerProvider(resource=resource)
-    span_processor = SimpleSpanProcessor(span_exporter=in_memory_span_exporter)
-    tracer_provider.add_span_processor(span_processor=span_processor)
+    tracer_provider = trace_sdk.TracerProvider()
+    tracer_provider.add_span_processor(SimpleSpanProcessor(in_memory_span_exporter))
     return tracer_provider
 
 
@@ -255,9 +289,7 @@ def instrument(
     in_memory_span_exporter: InMemorySpanExporter,
 ) -> Generator[None, None, None]:
     LangChainInstrumentor().instrument(tracer_provider=tracer_provider)
-    OpenAIInstrumentor().instrument(tracer_provider=tracer_provider)
     yield
-    OpenAIInstrumentor().uninstrument()
     LangChainInstrumentor().uninstrument()
     in_memory_span_exporter.clear()
 
@@ -329,6 +361,8 @@ LLM_INVOCATION_PARAMETERS = SpanAttributes.LLM_INVOCATION_PARAMETERS
 LLM_MODEL_NAME = SpanAttributes.LLM_MODEL_NAME
 LLM_OUTPUT_MESSAGES = SpanAttributes.LLM_OUTPUT_MESSAGES
 LLM_PROMPTS = SpanAttributes.LLM_PROMPTS
+LLM_PROMPT_TEMPLATE = SpanAttributes.LLM_PROMPT_TEMPLATE
+LLM_PROMPT_TEMPLATE_VARIABLES = SpanAttributes.LLM_PROMPT_TEMPLATE_VARIABLES
 LLM_TOKEN_COUNT_COMPLETION = SpanAttributes.LLM_TOKEN_COUNT_COMPLETION
 LLM_TOKEN_COUNT_PROMPT = SpanAttributes.LLM_TOKEN_COUNT_PROMPT
 LLM_TOKEN_COUNT_TOTAL = SpanAttributes.LLM_TOKEN_COUNT_TOTAL