+++ python/instrumentation/openinference-instrumentation-openai/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.1.1](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-openai-v0.1.0...python-openinference-instrumentation-openai-v0.1.1) (2024-02-08)
+
+
+### Bug Fixes
+
+* **python:** streaming with_raw_response in openai 1.8.0 ([#181](https://github.com/Arize-ai/openinference/issues/181)) ([6c8fcf0](https://github.com/Arize-ai/openinference/commit/6c8fcf0cf10aaacb3db777e906741ea2ea3496ac))
+
 ## 0.1.0 (2024-01-11)